group-service/requirements.txt
flask
requests
flask-session
PyJWT
